requirements.txt
flask==3.1.2
flask-sqlalchemy==3.1.1
flask-socketio==5.5.1
eventlet==0.39.0
requests==2.32.5
gunicorn==23.0.0
pytest==9.0.2
celery==5.6.2
redis==7.1.0
prometheus_client==0.24.1
python-dotenv==1.2.1
ruff==0.14.13
mypy==1.19.1
types-requests==2.32.4.20260107